Lagt til swagger tagger igjen

+++ src/models/item.py
@@ -3,6 +3,6 @@ from pydantic import BaseModel
 
 class Item(BaseModel):
     name: str = ""
-    price: float = 0.0
+    price: int = 0
     description: str = ""
     pictures: list[str] = [""]
